+++ bin/distribute_cli.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:args/command_runner.dart';
 import 'package:distribute_cli/builder_command.dart';
+import 'package:distribute_cli/create_command.dart';
 import 'package:distribute_cli/initializer_command.dart';
 import 'package:distribute_cli/logger.dart';
 
@@ -34,21 +35,20 @@ import 'package:distribute_cli/runner_command.dart';
 /// dart distribute_cli.dart build --config_path=config.env -v
 /// ```
 void main(List<String> args) async {
-  final runner = CommandRunner(
-      'distribute', 'Run commands to distribute your app packages.');
+  final runner = CommandRunner('distribute', 'Run commands to distribute your app packages.');
   final logs = File("distribution.log");
   if (await logs.exists()) {
     await logs.delete(recursive: true).catchError((value) => value);
   }
 
-  runner.argParser.addFlag("verbose",
-      abbr: 'v', defaultsTo: false, help: "Enable verbose output.");
-  runner.argParser.addOption("config",
-      defaultsTo: "distribution.yaml", help: "Path to the configuration file.");
+  runner.argParser.addFlag("verbose", abbr: 'v', defaultsTo: false, help: "Enable verbose output.");
+  runner.argParser.addOption("config", defaultsTo: "distribution.yaml", help: "Path to the configuration file.");
   runner.addCommand(InitializerCommand());
   runner.addCommand(BuilderCommand());
   runner.addCommand(PublisherCommand());
   runner.addCommand(RunnerCommand());
+  runner.addCommand(CreateCommand());
+  
   runner.run(args).catchError((e, s) {
     final logger = ColorizeLogger(true);
     logger.logError("${e.toString()}\n${s.toString()}");