+++ lib/app_builder/custom/command.dart
@@ -11,28 +11,20 @@ import '../../command.dart';
 /// tailored to their requirements for building the application.
 class Command extends Commander {
   /// Description of the command.
-  ///
-  /// Provides a brief explanation of what the command does.
   @override
   String get description => "Build a custom application by selecting specific configurations and options tailored to your requirements.";
 
   /// Name of the command.
-  ///
-  /// This is the identifier used to invoke the command.
   @override
   String get name => "custom";
 
   /// Argument parser for the command.
-  ///
-  /// Defines the arguments and options available for this command.
   @override
   ArgParser get argParser => Arguments.parser;
 
   /// Executes the command.
   ///
-  /// - Parses the arguments.
-  /// - Builds the application using the [AppBuilder].
-  ///
+  /// Parses the arguments and builds the application using the provided environment.
   /// Returns a [Future] that completes with the exit code of the build process.
   @override
   Future? run() => Arguments.fromArgResults(argResults!).build(Platform.environment);