+++ lib/parsers/task_arguments.dart
@@ -2,7 +2,7 @@ import 'job_arguments.dart';
 
 /// Represents a task in the configuration.
 ///
-/// A `Task` consists of a name, an optional key, an optional description,
+/// A [Task] consists of a name, an optional key, an optional description,
 /// and a list of jobs associated with the task.
 class Task {
   /// The name of the task.
@@ -13,17 +13,20 @@ class Task {
 
   /// The description of the task (optional).
   final String? description;
+
+  /// The list of workflow names associated with the task (optional).
   final List<String>? workflows;
 
   /// The list of jobs associated with the task.
   final List<Job> jobs;
 
-  /// Creates a new `Task` instance.
+  /// Creates a new [Task] instance.
   ///
   /// [name] is the name of the task.
   /// [jobs] is the list of jobs associated with the task.
   /// [key] is the unique key of the task (optional).
   /// [description] is the description of the task (optional).
+  /// [workflows] is the list of workflow names (optional).
   Task({
     required this.name,
     required this.jobs,
@@ -32,7 +35,7 @@ class Task {
     this.description,
   });
 
-  /// Converts the `Task` instance to a JSON object.
+  /// Converts the [Task] instance to a JSON object.
   Map<String, dynamic> toJson() => {
         "name": name,
         "key": key,